Extract homepage mission into new module
